plantaatoin

+++ src/abstraction/eq2Producteur2/Producteur2_Plantation.java
@@ -33,7 +33,7 @@ public abstract class Producteur2_Plantation extends Producteur2_MasseSalariale 
 	protected double rend_pest_MQ = 0.85;
 	protected double rend_pest_HQ = 0.80;
 	protected double rend_no_pest_HQ = 0.72; 
-	protected HashMap <Feve, HashMap< Double, Double> > plantation = new HashMap<>();	
+	protected HashMap <Feve, HashMap< Integer, Integer> > plantation = new HashMap<>();	
 	protected double start_years = 2024.0;
 	protected Journal journalPlantation;
